Rework parameter passing to decorator methods.
Rework parameter passing from
(result,router=0,decorator_helper,resrange,...) to more compresehnsive
version (session,decorator_helper). The session contains or should
contain all the parameters needed but there is a certain indirection
because session is being defined later in ulg.py. The rationale is that
even thouhg ulg.py can no be imported into files containing decorators
it is stable part of the ULG and therefore it might be used for building
decorators based on classes defined there (even though the classes are
not directly accessible, which can be overcomed by passing instances or
using decorator_helper to do needed class instantiation.

Modules for BIRD as well as Cisco has been changed to reflect this
changes.