Bugfix listing with regard to line delimiters.
Bugfix situation when listing slice is cut from text in session result
file it may happen that multi-line result is cut apart on listing
boundary which causes errors in table generation function of sh bgp
ipv4/6 unicast neigh <ip> received-routes and simmilar (handled by the
same shared code) commands. The errors actually means that there is
incomplete records because the cut part is missing (it might be
beginning of the first record or the end of the last record).

Also URL generation for modal windows was simplified.