Speed up BIRD BGP graph generation.
Refactor BIRD BGP graph generation by means of finishHook and arbitrary
data in sessions. FinishHook is used to set off result parsing and the
preprocessed list of ASes in paths is saved along with AS names resolved
from whois in the session data. In the display phase it is only
transformed into nodes and passed to graphviz.