Refactor graph generation and cover error path
Remove constants out of code and add error message to output when there
are no data produced by the show bgp ipv4/6 <prefix> command parser (or
BIRD equivalent - show route <prefix> all) and therefore no graph
can be generated. The error occures in situation when wrong prefix is
on input or the prefix has no paths in the routing table.