Add support for CommonSelectionParameter
There are problems with page size and page loading speed when there
are more routers with several peers and several BGP related commands
because selection commands multiplicate the peer tables in the resulting
javascript placed into the resulting page. This is not effective enough and
therefore sligh rewrite is needed.

Introduce new type of SelectionParameter - CommonSelectionParameter that
has its ID and it is placed to a JS variable in the beginning of the page
or function that uses the values inside only once. Then construction of the
select boxes uses this variables to feed needed values according to the
ID of the content.